fix(revit): Moved MeshBuilderHelper class to it's own file
